Include xtask into common, not the way around
