+++ alvr/common/Cargo.toml
@@ -32,5 +32,4 @@ widestring = '^0.4.2'
 winreg = '^0.7.0'
 
 [build-dependencies]
-toml = '^0.5.6'
-regex = '^1.3.9'
+alvr_xtask = { path = '../xtask' }
